move Book and Shelf types/interface to seperate file
